+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Set debugging off and ensure any existing breakpoints are disabled at extension activation
 * Fix toggling debugging to always send the `debug`/`nodebug` messages after enabling/disabling the breakpoints
 * Fix call hierarchy "Show Incoming Calls" and "Show Outgoing Calls" context menu items
+* Fix "Go to Symbol in Editor..." and "Go to Symbol in Workspace..." support to parse multi-line predicate scope directives
 
 ## [0.52.0]